restructuring every compontent hast now its own folder hope that makes the structure clearer
