+++ src/components/ResultList/ResultList.module.css
