+++ src/components/ResultList/ResultListEntry.module.css
