+++ src/components/Searchbar/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {ExampleButton, SymbolButton} from './Buttons';
-import {Store} from '../store/Store';
-import {triggerSearchAction, updateInputTextAction} from '../store/Actions';
+import {Store} from '../../store/Store';
+import {triggerSearchAction, updateInputTextAction} from '../../store/Actions';
 import styles from './SearchBar.module.css';
 
 /*
